requirements.txt
boto3==1.9.41
botocore==1.12.41
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
future==0.17.1
httplib2==0.12.0
idna==2.7
jmespath==0.9.3
oauth2==1.9.0.post1
oauthlib==2.1.0
python-dateutil==2.7.5
python-twitter==3.5
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
